+++ docs/release-notes.rst
@@ -29,6 +29,7 @@ Version 2.2.5, 16 July 2019
 * Add headers to discover texts (`ACUMOS-2992 <https://jira.acumos.org/browse/ACUMOS-2992>`_)
 * Add kernel type CC (`ACUMOS-3147 <https://jira.acumos.org/browse/ACUMOS-3147>`_)
 * Add kernel type JS (`ACUMOS-3148 <https://jira.acumos.org/browse/ACUMOS-3148>`_)
+* Update default catalog names (`ACUMOS-3209 <https://jira.acumos.org/browse/ACUMOS-3209>`_)
 
 Version 2.2.4, 23 May 2019
 --------------------------